requirements.txt
streamlit
pillow
tkintertable
pandas